Note that there are two types of alert the ORAC-DR pipeline should send to the eSTAR agent. An 'update' alert and an 'observation' alert.
An update alert is currently ignored by the agent, but should come at
the end of each observation frame, a 'observation' alert comes at the
end of all processing pointing to the final data products. For now
lets hardwire the sending of both alerts, after all we've just finished
a frame, and then (since the observation currently only consists of
a single frame) we need an update message to say that frame is done and
then immediately an observation message to say the observation is
done.

svn path=/trunk/; revision=4886