Version information comes from git version file.
version.sh is a new file to create a git version file containing the branch name, the
commit ID, and the date of last commit. This information is used by ORAC::Version to
obtain version information, which is then stored in the PIPEVERS header by ORAC::BaseNDF.
This commit requires a similar patch in Starlink::Versions.