lib: collate_headers now also updates RECIPE header
Modified ORAC::Recipe such that the uhdr() ORAC_DR_RECIPE header
is updated with the value that is in use by the pipeline.

Modified collate_headers so that it reads ORAC_DR_RECIPE and
translates that back to the native FITS header form. This ensures
that products from the pipeline list the recipe that was used
to generate them rather than the recipe that was requested in
the raw data. Important for CADC processing where modified
recipes (with suffices) are used by default.