SCUBA2: add a dark frame to list of input file if necessary
For the scan QL pipeline, darks will have been processed separately and stored in the calibration system. Retrieve the nearest dark frame and add it to the list of files passed to qlmakemap. A check is made on the existence of the SHUTTER keyword in the subheaders - if it is present then the file list will already contain darks and thus it is not necessary to pick one from the calibration system.