SCUBA2: Refactor code to work properly with live pipelines
For online pipelines, the data appear at intervals, so processing is
deferred until all data has been taken. A copy is made of the raw data
to ensure it remains available for processing. The current and
previous responsivity solutions are now stored as NDF extensions in
the flatfield solution under .MORE.SMURF.RESP_NEW and
.MORE.SMURF.RESP_OLD.