SCUBA2: retrieve a suitable dark frame if necessary
Add a check to see if the current list of input files contains any
dark frames. Use the presence of the SHUTTER keyword in the current
Frame subheaders to indicate the offline pipeline. If SHUTTER is not
present, then the input files do not include dark frames (as in the
online pipeline) so one must be retrieved from the calibration system.