Use ORAC_SURVEY instead of SURVEY.
Some older ACSIS data has a blank SURVEY header. A modification to Astro::FITS::HdrTrans was made that creates the
generic "SURVEY" header, looking at the PROJECT header if the SURVEY header is blank. This patch uses the generic header
so, if the SURVEY header is fixed for old data, we only have to patch one place.