SCUBA2: throw error in case of makemap failure, update POD
Trap the general case where makemap exits with bad status and throw an
error to abort the current recipe. Mark the frame as bad on
failure. Also always pass in an ORAC::TempFile to _GET_MAKEMAP_CONFIG_
to allow internal consistency checks to re-write a config file.

Update the POD to bring it into line with other primitives.