SCUBA2: update POD, deal with short data files
The iterative map maker apodizes the time stream before filtering. If
the user attempts to process files which contain fewer samples than 4
times the apodization length, the map will be blank and makemap will
exit with an error. The new code checks for the length of the data
file(s) and adjusts the apod parameter accordingly. Since the default
value of apod is 1000, This situation only arises for very short
science observations (<20-40 sec) of bright sources (e.g. planets).

Bring POD into line with recent updates.