SCUBA2: assess image before attempting to find offsets, add warning if it failed
Use the new _ASSESS_IMAGE_ primitive instead of _FIND_POINT_SOURCES_
to determine whether or not the input image has a well-defined point
source which can be used to derive pointing offsets. Since we only
want to do this for point sources, check the explicit return value of
the POINT parameter. A warning is issued if no source could be
identified.

Some code has been rejigged - the copy of the group file is made
before writing the flag file; there is no need to initialize the log
file unless results exist to be logged.

POD updated to reflect use of new primitive.