heterodyne: Save and restore DR recipe header.
This patch was made because when the -recsuffix command-line option is
used, the new recipe name is stored in the $Frm->uhdr. However, when
copying raw data and updating the Frm object, the original recipe name
overwrites the new one. If we keep the new one and update the Frm object,
we can restore the recipe name to what it should be.