SCUBA2: retrieve bad bolometer masks and store files listed in OUTFILES
Attempt to retrieve bad bolo masks for each subarray and pass that
list to flatfield. Files listed in the OUTFILES parameter are stored
in the Frame object instead of using the inout method. Note that this
will disrupt the naming scheme since the subscan numbers in the
filenames will now not correspond to those in the files.