SCUBA2: new primitive for processing SCAN data
This primitive is designed to be used for the summit pipeline only and
allows the user to delay the map-making step until a certain
proportion of the observation has completed. The data are flatfielded
in the meantime. Other processing tasks may be carried out as deemed
necessary.

The rebinning map-maker is also supported should the user wish to
compare, though in this case there is nothing to gain from waiting so
the PERCENT_CMP parameter is not used.