general: No longer subclass flatindex method for NACO, Michelle and IRIS2
These instruments had an implementation of flatindex that used the
name index.flat for the index file but picked up a rules file
that depended on observation mode. This doesn't seem to be
right since the rules files are different and so the columns
in the shared index file won't match. Synchronize with other
imaging spectrometers by using ImagSpec.pm directly and
using index.flat_sp and index.flat_im