SCUBA2: a generic primitive to create a Frame image
This is a new generic primitive to create a Frame image. It can cope
with the DREAM/STARE and SCAN pipelines, whether quick-look or not.

However, it is rather basic and relies on sensible defaults for the
other primitives called within since it is not practical to support
every possible argument. It is designed to be used in the QL pipeline
for non-science observations.