general: Disallow processing on an NFS mounted disk
Be a bit draconian and disallow any attempt to write to an
ORAC_DATA_OUT that appears to be mounted over NFS. We don't
want to encourage this sort of thing and it usually causes
pain and suffering to allow it. A warning might work but give
an error a try first.