SCUBA2: new primitive for steering the processing in SUMMIT recipes
Currently sets new Frames for SCAN data to `bad' if a new Group image
has been created and sets the NOCALIB flag to true so the data aren't
calibrated though this is less important as they will at least be
flatfielded. These frames are ignored to prevent mosaicking errors
when combining 2-d images and unprocessed 3-d timeseries data. This is
easier than checking the dimensionality of each file.