SCUBA2: don't bother processing data after a new image has been made
For the summit pipeline there is no need to pick up processing of any
new data which is taken after the map has been made. First, set the
OBSEND flag to true so that a new Group file will be made (on the
understanding that the WAIT parameter is specified for
_MAKE_MOSAIC_GROUP_). Second, set a Group uhdr entry to denote that a
new Group image will have been made. Make sure this header is unset
for times when the data are just to be flatfielded, and that they are
marked as bad so a group mosaicking step doesn't try to including
these timeseries data.