scuba2: Create a file on init to help the JOS know where we are writing
The JOS needs to read data files from the reduced data directory
but it is difficult to guess where this may be if the DR
moves around from machine to machine (given the new mount point
layout at JCMT). When oracdr_scuba2x init script is run at the
summit we now write a file to /jac_sw/oracdr-locations to indicate
where ORAC_DATA_OUT is.