core: Refactor list data detection with pattern for scuba-2
The logic for locating files using pattern_from_bits was in
two places (orac_loop_list and link_and_read) so this patch
moves that into a single subroutine.

That subroutine is then also extended to allow for the returned
pattern to refer to a flag file rather than a data file. This
is required for SCUBA-2 where the flag files are in ORAC_DATA_IN
but the data files themselves are not in a sub-directory of
ORAC_DATA_IN.

Along with a change to ORAC::Frame::SCUBA2 to return pattern
of the .ok file rather than pattern of the raw data files, the
SCUBA-2 pipeline can now work in -list mode and in -batch
mode (which triggers -list mode).