JSA: Add option to flag non-zero quality pixels as bad before tagging as reduced product.
This patch was done because CADC doesn't store quality fields. In reduced
ACSIS cubes, a tile border is created, which overlaps in neighbouring
tiles. A quality flag is set denoting which pixels are in the tile border.
If these pixels are not flagged as bad before the data are sent to CADC,
people downloading the cubes will not know that these pixels are
duplicated between tiles, and errors in combining the results would
probably occur.