SCUBA2: deal with Frames which contain only darks
Now that Frames consisting solely of darks are not processed by a
separate recipe, the top-level primitives have to deal with this
case. Check the status of the DARKONLY flag and proceed only if not
set. In some cases the dark data are copied to the output directory
where they can be picked up later if necessary.