lib: Add ORAC::Error::TermProcessing and orac_term()
Add a new exception and related command to easily be able to issue
an error message and terminate the current recipe processing.
orac_throw is not suitable because it forces ORAC::Error::FatalError
and that is deemed to be really fatal.

Will have to change all orac_throw calls in primitives to orac_term
once the execution infrastructure is modified to understand the
difference.