SCUBA2: shift images before creating cube to put the source at 0,0
The pointing and focus task at the telescope assumes the source is at
(0,0), which is not necessarily true at least for now. So find the
peak position for the middle focus position (assumed to be closest to
the actual best-fit focus value) and shift all the input images to get
the source to lie at 0,0.