heterodyne: Supply full bounds when creating NDF section. Trim bad pixels.
NDF sections can't easily handle unbounded WCS sections, as if you have a
WCS axis whose units run opposite to the pixel axis, you'll end up getting
the reverse of what you actually want. We define the full range in WCS
coordinates to remove any possible ambiguities as to what we want.

Also trim bad pixels from the output NDF. This is so that if our input NDF
runs from, say, -50 to 100, and we want to "trim" to 0 to 200, we end up
with an NDF whose bounds are 0 to 100 and not 0 to 200.