SCUBA2: Let SCUBA-2 QL drive a display natively
For dream/stare observing the QL can't display fast enough so we
need to use oracdr_monitor. For scan mode we have loads of time
so to simplify startup and display reliability enable the display
in the S-2 QL pipeline.

If DREAM/STARE is ever enabled we need to reconsider.