SCUBA2: method for calculating nefd is now determined automatically, write logfile
The choice of which method to use is now determined from the
observation type: dark and skydip obs will have noise data from which
the NEP statistics are derived and converted to NEFD estimates for the
current airmass and zenith. Other observation types will use the
variance and EXP_TIME components of the current Group or Frame
image. For these observations, the variance and exposure time are
derived only for the area of the map specified by the FITS header to
avoid being biassed by the poorly-sampled map edge.

The user can now specify whether they want to use the median or mean
value for estimating NEP, EXP_TIME etc. The mode should also be an
option, but histat has problems with the range of NEPs for the current
850 um array.

A logfile is now written for both cases.