lib: ORAC::TempFile now works with OPEN => 0
Using OPEN to close the file has never worked because the constructor
required an open filehandle so that it could close it but OPEN is used
by File::Temp to close the file before ORAC::TempFile gets anything.
This worked with a single false argument but becamse obvious when
attempting to try "SUFFIX => '.sdf'" as well.