general: Strip leading whitespace before writing AST Region to file.
AST can't handle strings longer than 255 characters (for the time being), and if
we have combined a large number (>~10) Regions to form a CmpRegion, lines in the
textual representation of that CmpRegion can be longer than this limit. If we
remove leading whitespace the lines come in under this limit.