SCUBA2: major overhaul (and simplification) of determining config file names
Support for the MYCONFIG option to _MAKEMAP_FRAME_ has been dropped in
favour of the simpler system of just passing the name of an
alternative config file to this primitive (e.g. CONFIG=$config).

The primitive checks that the given file exists, using it as is if
given an absolute path, otherwise checking the current working
directory, then ORAC_DATA_OUT and ORAC_DATA_CAL in that order.