SCUBA2: do not check map size in the QL pipeline
In the QL, only 30-60s of data is processed at a time, and it's not
possible to scan an area large enough to create an oversize map in
that time. Even though this check will not take much time, we might as
well claim as much as we can to ensure the QL can keep up.