SCUBA2: new primitive to check map bounds before making an image
This primitive replaces code repeated in _MAKE_MOSAIC_ and
_MAKEMAP_FRAME_ to check the size of the output image before creating
it to ensure that it remained below 2 GB in size.

The user must specify both an input filename which contains the list
of files to process, and the task to use to check the mapsize: SMURF
makemap for raw data, KAPPA wcsmosaic for images.

The new bounds are returned to the caller via the entries UBOUND and
LBOUND in the primitive hash.