SCUBA2: update aperture method for noise estimate
Now only use 4 (for small images) or 8 apertures (for larger images)
to estimate the noise. The source position is avoided. A simple check
is made on the results to look for high outliers which might indicate
the presence of another source (tested with DG Tau data which has a
second source about an arcmin to the SW). These are removed from the
data before the noise is calculated as the mean of all the apertures.