SCUBA2: check that output map exists before storing
Under certain conditions (such as a loss of heater tracking), raw data
files end up on disk which smurf deals with in a way that means the
exit status is good even though no output image is created. This
primitive previously only checked for good status as an indicator that
a successful output file was created but now checks that the expected
image actually exists in ORAC_DATA_OUT before storing the file in the
current Frame.