SCUBA2: change the criteria for processing data into images
In order to distinguish unprocessed data, the Frame isgood attribute
is set to -1. This guarantees that the data still appear to be good
when a simple truth test is made.

The change in the TCS_INDEX is now the dominant criteria for making a
map; the percentage completion parameter is only checked for
observations in which the TCS_INDEX does not change (ie the scan
pattern is only completed once during the observation).

The Group uhdr LAST_INDEX is updated to the current TCS_INDEX if a map
is made.