SCUBA2: use recipe parameter for aperture radius, default to 2*FWHM
The BEAMEQUIV and ARCSEC calibrations use aperture photometry to
derive an FCF. Previously this was fixed at 30 arcsec but now the
default depends on the wavelength. The recipe parameter
APERTURE_RADIUS may be used to override the default. Note that the
SCUBA-2 pipeline does not yet support this recipe parameter; however,
several PICARD recipes do.