SCUBA2: do not issue a warning if sky subtraction has been done
For the summit pipeline there is no point constantly warning that sky
subtraction has been done when the primitive is actually being skipped
because there is no image to actually remove a background from.

Also the warning is meaningless when the primitive is being used to
remove a background rather than the `sky'.