SCUBA2: crop images before adding to focus cube
The image for each focus position is now cropped to 90x90 arcsec to
improve the display and perhaps ensure more accurate peak fluxes to
solve for the best-fit focus position. Cropping the images required a
smaller box to be passed to _FIT_BEAM_ in order to guarantee the
minimum number of good pixels.

The display format is now arcsec offsets.