SCUBA2: add meta_file method to return path to the .meta file and use it
The summit pipeline needs to know what flag files are to be
written. The DA system writes a .meta file in ORAC_DATA_IN with the
name of the .ok files for which data have just been written. That way
the pipeline is not kept waiting for data that don't exist if one
subarray stops working.