SCUBA2: refactor logic for map-making
A new Grp uhdr entry called SCANCOMP is set if a map is made on
completion of the scan pattern.

Note that if a map has been made as a result of the scan pattern being
completed, the percentage-completion check is ignored. (Note that the
reverse is already true - if the percentage completion parameter is
set, there is no check on whether the scan pattern has been
completed. This is handled in _SUMMIT_STEER_.)