SCUBA2: change scaling for display purposes, add obsnum to image title
The display scaling is now defined in terms of the median noise
value. The new parameter THRESH controls the maximum value to display
(both for the image and the histogram) and is a multiplier for the
median. The old CLIP parameter has been removed.

The effective NEP is now written to the bolometer noise file rather
than its own separate file. The number of bolometers used in the noise
calculation and the modal noise value are also written.

The current observation number is written in to the title of the noise
image so it appears on the display.

The internal documentation has been updated to reflect these changes.