SCUBA2: delay processing by 1 file if percentage completion satisfied
In the case where the observation consists of multiple passes of the
scan pattern it's possible that the percentage completion criteria
will be picked up first. If this happens, the remaining data would be
ignored which is clearly not desireable if the observation is really
only 10% done!

Now the primitive sets a flag which causes the pipeline to wait for
one more file to arrive before making a decision. If the TCS_INDEX has
changed for that new file then use the scan-pattern-completion
criteria for making a map. Otherwise use the percentage completion
criteria (and ignore subsequent data).