SCUBA2: add BEST_PERCENT argument to create a mask using the given fraction of bolometers
By default the best 50 % will be used (corresponding to the
median). This argument overrides the RELAX argument.

The value of the NEP used to threshold the data and create the mask is
written to the FITS header of the mask, as is the number of bolometers
in the mask and the number of good bolometers in the original data.

The recipe parameter BESTBOL_PERCENT is supported to allow offline
adjustment of this parameter.