wfcam: Do not check for negative mean for skyflats
From Andy Adamson:

Completing an action on me from a schedule meeting a couple of weeks
back: the problem was that when UKIDSS GPS and LAS use a 1-s-exposure
SKYFLAT in Z, Y (and maybe J) as a way to flush the array between filter
changes, the mean count can be negative and the DR stops. Looking at the
QUICK_LOOK recipe, which is what's used in these cases, it calls
_IMAGING_HELLO_, which for WFCAM includes a _CHECK_DATA_INTEGRITY_,
which in turn checks for zero and negative mean and channel edges.

Since flush frames are just thrown away anyway, this check is
superfluous in their case. The association between QUICK_LOOK and
skyflat flushes is all over the place in UKIDSS, so it wouldn't be easy
to get out of by adding a new recipe. The end of night skyflats, which
are real flats, use a different recipe (which presumably still calls
_IMAGING_HELLO_), but the reduction of those frames basically only
reports the count rate (doesn't add them into groups for example).

So if I'm right about this, I think the best option is to have
$ORAC_DIR/primitives/imaging/WFCAM/_CHECK_DATA_INTEGRITY_ not do the
"mean<0" check for OBSTYPE equal to SKYFLAT.

All we would lose in doing this would be the check for *negative* counts
in the bona fide skyflat frames at the end of the night; but that's ok
because the TSS is monitoring the counts in those flats anyway.
(cherry picked from commit 0eaa80c506bbf2a81ea20bd8a22ac9439340d0c7)