SCUBA2: new primitive for the QL pipeline to determine the average noise properties
The average noise properties are derived from the average of the
individual noise images. Those images are stacked with stackframes
then collapsed to obtain the mean, which is analyzed in the same way
as the individual images. The image title for the display informs the
user that it is the mean noise image.