picard: All files given to picard must be separate frame objects
framegroupkeys had been implemented incorrectly for picard. It
had assumed that returning an empty list meant they would be
separate when in fact it forced them to be a single frame object.

Now framegroup() is subclassed.

With this change picard again puts each file in a Frame object
and all the Frame objects in a single Group object.