PICARD: ensure a minimum map size if the requested size is too small
For calculating the image-plane noise (in _CALC_FLUX_), the image must
not be cropped too small - twice the diameter of the aperture is the
minimum size. Only proceeds if the APERTURE_RADIUS recipe parameter is
defined (which it won't be if this primitive is called in the
CROP_IMAGES recipe).

Still doesn't deal with input images that are too small.