heterodyne: Change naming scheme for raw data copy
Rather than appending _raw$i which is a counter corresponding
to the file number in the frame, we insert "raw" before the last group of
numbers.

This is required so that the name more closely resembles the original.
Ordinarily it does not matter but for hybrid mode observations
with differing subsystem numbers we need a way of working out
the original raw data file in the absence of provenance.

Provenance is missing because we do not add a root provenance
entry when we NDFCOPY the file. We should add an entry from
the raw file and store OBSIDSS. The reason we don't is because
currently this primitive only optionally copies the file. Sometimes
it hard links the file for speed but we should not be modifying
the raw data even if it is local.

The other way we could fix things is to always copy the data
and sacrifice some startup speed.