PICARD: align NDF components before adding, rejig params handling, deal with ACSIS/SCUBA-2 differences
Use the updated version of _ALIGN_ARRAYS_ to align the exp_time and
other NDF components.

The handling of the params argument for wcsalign/wcsmosaic has been
changed to allow easy use when calling _ALIGN_ARRAYS_.

ACSIS doesn't have WEIGHTS but does have EFF_TIME so add code to deal
with that difference. The TSYS NDF component is not yet dealt with
here - it will require a different estimator to simply adding.