SCUBA2: return smoothed version of PSF if requested
Add SMOOTH and SMOOTH_FWHM parameters which control the smoothing of
the PSF. If SMOOTH is true, then the PSF is smoothed with a gaussian
with a FWHM equal to SMOOTH_FWHM which is subtracted from the original
PSF image. This smoothed-and-subtracted version is returned to the
calling primitive.

Also update internal documentation and add some user feedback.