SCUBA2: no need to undo convolve normalization, only smooth science image
The SMOOTH parameter has been turned into a flag to denote whether the
image should be smoothed and have that smoothed version subtracted
from it. The size of the Gaussian for smoothing is now controlled by
the SMOOTH_FWHM parameter.

Now only the science image is smoothed - if the PSF is to be created
then the smoothed version is returned from that primitive.

Use the new norm parameter for convolved to prevent
normalization.