Actually return bad exit status if a recipe failed to run
ORAC-DR tries not to stop the pipeline for any reason. This is
a problem when you want to know that some recipes failed. Now
on exit the pipeline will report the number of failed recipes
and return a non-zero exit status if that number is non-zero.

Also revamp the exception catching in the main oracdr and picard
programs to reduce code duplication and to also force those
cases to exit with bad status.