General: skip lines beginning with # in read_file_list
Although this routine was written to read files created by the
OUTFILES parameter in SMURF, it can be used to read any text file,
returning an array of lines. The caller can then parse the lines
accordingly. To enhance its general usefulness, skip lines that are
conventionally comments - ie those beginning with #.