scuba2: Tweak default dimmconfigs
This is a bit of a quick hack so that we can reprocess data at CADC.
The config names have changed so that bright_compact is used for
calibrators, blank_field is used for blank fields and we use
dimmconfig.lis for everything else by default.

Will need to revisit this patch.