lib: Add ability to terminate early without hiding the error
orac_term stops the recipe early but is not treated as an error
condition. orac_throw aborts the pipeline. Add orac_termerr
to terminate the recipe early but with an error state so that
the pipeline will track the failure. This can be used by the
recipe to indicate that the recipe had a problem but it was
handled.