SCUBA2: generalize call to makemap or wcsmosaic, return -1 if task failed
Since the wcsmosaic and makemap calls to retrieve the bounds are
identical, the code which runs the task can be generalized.

Now check the return status of the task and return -1 in the primiitve
hash if the status was bad. Note that no error is issued here - it is
up to the calling primitive to deal with the outcome.