SCUBA2: rejig handling of failed map-making attempts, do not terminate recipe if error occurs for last member
If an error occurs processing the last member of a Grp, immediately
terminating the recipe is a bad idea because no Group file will be
created, even if all the previous Frames were good. Therefore, only
terminate the recipe if an error occurs processing anything other than
the last Frame. In that case, just issue an error and continue.