acsis: Try to get consistent grouping for DR and database
The ORAC-DR grouping algorithm uses different metadata depending
on whether there is information in the FITS header or from the
WCS component. Headers constructed from a database query were
giving different results for header grouping because of an
inconsistent use of sprintf. We are not allowed to change
DR file behaviour so this patch tries to get the database
form to agree.