SCUBA2: add GAUSSIAN parameter to allow non-Gaussian fits
Beamfit recently added the gauss parameter which allows the radial
fall off parameter to be varied rather than force a Gaussian. PSF has
had this capability for a long time, so this commit means that the
differences between _FIT_BEAM_ and _FIT_PSF_ are reduced. I see a
single merged primitive on the horizon...