PICARD will now attempt to use an instrument-based search path
Now PICARD will check the instruments of the files it has been
given and if they are all the same it will attempt to guess
the corresponding ORAC_INSTRUMENT value. $Cal will then be
recalculated as if it was of that instrument and the
primitive search path is adjusted so that PICARD comes in
front of the normal instrument paths.

This may have broken things so keep an eye out. Note that currently
the frameclass is not switched but it would be trivial to make
that change. We still want to aim for primitives that are being
used in a PICARD context to use the generic translated FITS
headers whenever possible.