SCUBA2: use _FIT_SOURCE_ and tidy up logic so it works as intended
It looks like I was expecting the `if ($x && $y)' test to mean if ($x
AND $y) is true, whereas it's actually if ($x is true AND $y is true)
which behaves the same when $x and $y are both non-zero, but
differently when $x and $y are both 0.

Also add support for recipe parameters FIT_GAUSSIAN, FIT_METHOD and
LOGFILE for use with PICARD recipes.