SCUBA2: do not delete the flag file if OBSEND is true
Now that pointing recipes use the matched filter, much of the
processing that was delayed until the end of the observation is
carried out as part of the processing. This means that no further
estimates of the pointing will be made and so any existing flag files
should not be deleted.