SCUBA2: only calculate pointing offsets if a new image was created in QL
The QL pipeline originally calculated pointing offsets on every pass
through the recipe if data existed. It's not necessary to do that any
more so only derive an estimate if a new image was created during the
current pass through the recipe.

Also allow the radial fall-off parameter to vary in the source fits
(appropriate for output from the matched filter).