SCUBA2: only calculate beam size if a new image was created in QL
The rationale bahind this change is that the beam size should be
measured on images from the map-maker, not anything which has been
processed with the matched filter. Therefore create this simple no-op
to prevent beam fitting during passes through the recipe with no
science data (e.g. at the end of a pointing obs).

Also print beam properties in green for easier on-screen viewing.