scuba2: At CADC we do not want to run the pointing recipe
So the _CADC variant currently points to the default scan
recipe. The pointing recipe will be removing background,
truncating to map area and possibly applying a matched
filter to optimise point source detection.

Pointing observations should be treated like any other science
data so as not to confuse the science archive.