SCUBA2: re-register images of bright compact sources
If the data have been processed using the `bright_compact' config file
then it should be safe to assume that there is a source in the field,
somewhere near the centre (20 pixels) which can be detected and its
offset from the nominal centre estimated.

Note that the images are shifted in place - the Frame files are
updated.

A further enhancement would be to run the matched-filter to locate the
source positions with greater accuracy, and for sources of low
signal-to-noise.