SCUBA2: allow a Frame object to be passed in
Since this primitive created a Frame object from the given files, it
makes sense to be able to pass in a Frame object directly. This will
allow future upgrades to make assessments of the images and remove
those with poor S/N or no source detection from the Frame before
coadding.