SCUBA2: use a temp file for the PSF if the output is not being stored
The thinking here is that when this primitive is called with STORE=1
it's part of a recipe that cleans up after itself, this getting rid of
any _mf and _psf files. If STORE=0 then just use a temp file for the
PSF and don't bother pushing it onto the intermediates array.