SCUBA2: try and find a reason for failure to determine FCF
For Frame objects, if the FCF calculation failed due to the source
position not being found, then examine the peak S/N in a 30x30 pixel
box to see if the source *could* have been detected. Set threshold at
5-sigma. If the S/N is high enough, but no source was detected then
the data are probably bad in some way. Assume that low S/N means that
the data are otherwise OK, and just not sensitive enough to allow the
source to be located.