SCUBA2: make copies of input files if desired, make it work with 1 file
The user has the option of requesting that the original images not be
shifted (wcsslide modifies in place), so copies of these files should
be used instead. These have the suffix "_reg".

Previously this primitive did not work if the Frame only contained 1
file since it was assumed to be called to shift multiple files at
once.