SCUBA2: rewrite of peak-finding algorithm to allow blind searches
_FIND_MAP_PEAK_ is now capable of a blind search of the image to
return the position of the maximum signal. It now creates a
signal-to-noise ratio image which is processed with the matched filter
(provided 5-sigma peaks exist). The position of the brightest peak in
the filtered image is then used as an initial guess for fitting a 2-d
profile (if the FITPEAK parameter is true). The fitted peak value and
position are returned.

The user may optionally provide an initial guess for the position, in
which case only a 2x2 arcmin^2 region about that position is used as
above.