SCUBA2: get sign right for dec shift, add INPLACE parameter, update POD
The RECPAR_REGISTER_NOREPLACE recipe parameter has been replaced by a
new argument called INPLACE - should be true for the pipeline, false
for PICARD. That is, files produced in the pipeline are automatically
shifted in place, whereas the original input files given to PICARD are
left untouched.

Get the sign right for the declination shift for sources below the
celestial equator.