SCUBA2: add capability to calculate the RMS
The NEPTYPE argument has been added to allow the user to choose
whether the RMS or effective NEP is to be calculated. The calculations
are very similar so it putting the two into a single primitive seems
like the right thing to do.

Oops this went into the rename commit as well...