SCUBA2: calculate mapping speed, write results to log file(s), set units
Noise statistics and NEPs are written to log.bolonoise.

The mapping speed is calculated for the best 300 to 1000 bolometers,
and, along with the corresponding RMS NEPs, is written to
log.mapspeed_SUB, where SUB is the subarray designation.

Read the NEP units directly and set them in the output NEP image.