SCUBA2: new primitive to create an NEP (hyper)cube as a function of heater and bias
It would be easy to make this create just a 3-d cube of NEP vs heater
or bias, but that will have to wait for a future update.

Currently always creates the 4-d hypercube and doesn't tell the user
much about what's going on. The code for creating the WCS can be
factored out into a separate primitive since it's almost identical to
that for _CREATE_FOCUS_CUBE_.