SCUBA2: recipe for processing NEP observations performed in engineering mode
This recipe calculates the noise and NEP for each heater and bias
setting (and each frequency range if appropriate) to produce an NEP
hypercube of the NEP as a function of bolometer and heater/bias
settings. In addition, images are produced of the RMS and effective
NEPs per subarray, and the array `mapping-speed' parameter is
calculated as a unction of the best N bolometers. Results are written
to logfiles.