heterodyne: Tweak frame creation to work with raw files
Raw files don't magically get their headers read as a side effect
of calling sync_headers so if this primitive is called with a raw
file (ie without copying the data) it was crashing. Now use
Astro::FITS::Header to read SUBSYSNR and force a readhdr()
when creating temporary frame object.