SCUBA2: deal with non-4d cases
Only create a 4-d hypercube if there are multiple values of both the
heater and bias settings. If one of these has only 1 setting then a
3-d cube will be created. In addition if both have only 1 setting,
then the output images will be copied to the `cube' names.

Still need to work on factoring out the WCS into a separate primitive,
but it works for now.