SCUBA2: new primitive to perform the mapping speed calculation
Loops over all files in all Frames in the Grp and calculated the
mapping speed for different numbers of bolometers. The range is
user-specifiable and may be overridden with recipe parameters.

The results are written to a log file, one per subarray.

This is code factored out of _CREATE_NEP_IMAGE_ and helps clarify the
procedure for processing NEP data.