SCUBA2: pass a position to _FIND_MAP_PEAK_
Passing in a position to _FIND_MAP_PEAK_ will alleviate some of the
issues with calculating FCFs for weak sources as the map will be
cropped to 2x2 arcmin^2.

Use the skyref position, unless the map has an offset coordinate
system in which case use (0,0). This should be OK as a first attempt
though it assumes that the source we want to calculate an FCF for is
at (or near) the SkyRef position, as it should be for the current
calibrator test. Things could get trickier for sources with
user-specified fluxes as there's no guarantee the SkyRef is the actual
source position.

Other things: add a bit of user feedback, mark failed FCF
determinations in red, and actually we don't need the set the
BEAMMATCH entry to 0 in the results hash.