SCUBA2: new recipe and primitives for adding fake sources to SCUBA-2 data
This new recipe allows the user to give a map to be added to the raw
timeseries data. This map is used to create a `fake map' with the same
pixel bounds as the SCUBA-2 map would have. The map may be regridded
to match the pixel scale with the output, and may be shifted on the
sky so that sources do not overlap.

Four recipe parameters are supported:
 * FAKEMAP_MAP: name of the input map
 * FAKEMAP_SCALE: scaling factor used by the map-maker
 * FAKEMAP_OFFSET: arcsec offsets to shift the input map
 * FAKEMAP_REGRID: flag for regridding map to same pixel scale

Note that the original map is not touched: a copy is made and all
changes are made on the copy.

If the _REGRID flag is given, the map will only be regridded if the
pixel scales differ by more than 1%.