SCUBA2: add more flexibility in creating reference images
The user can now specify the desired pixel bounds, a flag to denote
whether or not WCS should be added, and provide an Astro::Coords
object to specify the coordinates of the reference image. Previously
the Group was always used to define filenames: now the current Frame
can be used as well.

Note these changes have only been tested for fakemap processing, not
DREAM/STARE as this primitive was originally written for.